requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

nltk>=3.7
rouge-score>=0.1.2
sacrebleu>=2.0.0
bert-score>=0.3.11

torch>=1.12.0
torchvision>=0.13.0

librosa>=0.9.0
pystoi>=0.3.3
soundfile>=0.10.0

scikit-learn>=1.1.0
scipy>=1.8.0
Pillow>=9.0.0
tqdm>=4.64.0
psutil>=5.9.0

unbabel-comet>=2.2.2  # optional